Reorganise sagas & reducers
