+++ app/redux/compass/compass-reducer.js
