+++ app/redux/current-user/current-user-saga.js
@@ -2,8 +2,8 @@ import { put, takeEvery, select, take, cancel, takeLatest } from 'redux-saga/eff
 import firebase from 'react-native-firebase';
 import { eventChannel } from 'redux-saga';
 
-import { CURRENT_USER_COLLECTION, GEO_POINTS_COLLECTION } from '../constants';
-import { calculateAgeFrom } from '../utils/date-utils';
+import { CURRENT_USER_COLLECTION, GEO_POINTS_COLLECTION } from '../../constants';
+import { calculateAgeFrom } from '../../utils/date-utils';
 
 export default function* currentUserSaga() {
   try {
