+++ app/redux/map-panel/map-panel-reducer.js
