+++ app/redux/map-panel/map-panel-saga-new.js
@@ -1,7 +1,7 @@
 import { call, take, put, actionChannel, cancel, takeLatest, select, throttle } from 'redux-saga/effects';
 import { buffers, delay } from 'redux-saga';
 
-import GeoUtils from '../utils/geo-utils';
+import GeoUtils from '../../utils/geo-utils';
 
 const mapPanelReplaceDelay = 250;
 // const mapPanelHideDelay = 400;
