+++ app/redux/map-view/mapview-my-visibility-saga.js
