+++ app/redux/map-view/mapview-saga.js
