+++ app/redux/upload-photos/upload-photos-reducer.js
